+++ ui/src/mixins/index.js
@@ -1,4 +1,5 @@
 import contextMixin from './context.js'
+import formMixin from './form.js'
 import generatorMixin from './generator.js'
 import passwordMixin from './password.js'
 import screenMixin from './screen.js'
@@ -6,6 +7,7 @@ import viewMixin from './view.js'
 
 export {
   contextMixin,
+  formMixin,
   generatorMixin,
   passwordMixin,
   screenMixin,