+++ CHANGELOG.md
@@ -9,6 +9,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 - `QasField` changed emit value when field is type number.
 
+### Fixed
+
+- Added cloned value for `field` in computed `children` inside `QasNestedFields`
+
+## 2.2.0 - 2021-05-26
+
+### Added
+- Added `formMixin`
+- Added new slot `custom-slot` inside `QasUploader`.
+- Added `readonly` and` maxFiles` properties in` QasField` for `QasUploader`.
+- Header added `Content-Disposition: Attachment` for all uploads made with `QasUploader`.
+- Added margin prop in `QasLabel`.
+
+### Changed
+- Adjustments on `QasCopy` icon size to respect design system.
+- Change the `QasAvatar` image prop, not running into asset function.
+- Change `QasAvatar` letter weight to bold.
+- CSS classes now have `!important` flag.
+
+### Removed
+- Remove `.q-pa-sm` actions slot in `QasCard`.
+- Remove require prop `QasLabel`.
+- Remove position absolute in `QasSettings`.
+
+### Fixed
+- `QasUploader` now support multiple uploads.
+- iconSize in `QasSettings`.
+- Fixed import of `filterObject` helper.
+
 ## 2.1.0 - 2021-05-15
 
 ### Added